+++ tests/test_init.py
@@ -7,7 +7,7 @@ import vigilo
 
 class TestInit(unittest.TestCase):
     def test_version(self) -> None:
-        self.assertEqual(vigilo.__version__, "0.3.0")
+        self.assertEqual(vigilo.__version__, "0.3.1")
 
 
 if __name__ == "__main__":